reformat fix setup of default ui font from file update debug messages fix datatype used for the pointsize in font-loader

+++ common/src/TrenchBroomApp.h
@@ -103,7 +103,7 @@ private:
 public:
     void reloadStyle(bool reloadFonts = false, bool reloadStyleSheets = true);
 
-    QFont loadFont(const std::filesystem::path &path, const size_t size);
+    QFont loadFont(const std::filesystem::path &path, const int size);
 
     const std::vector<std::filesystem::path> &recentDocuments() const;
